feat(console): @kagami/console-api 契约包，管理台查询路由契约化（#279 PR4）
- 新包 @kagami/console-api：app-log / napcat-event / napcat-group-message 三个 schema 文件
  自 shared 整体迁入；llm-chat.ts 的历史查询段（LlmChatCallStatus~DetailResponse）抽出为
  llm-chat-call.ts（该段只依赖 wire 基元，与 LLM 负载核心 schema 无耦合）
- 五条路由进契约，/llm-chat-call/:id 用 PR1 的 params 通道（preprocess 字符串转 int 语义保留）
- console 四个 handler 换 registerJsonRoute(contract)，既有 13 个行为测试期望值零改动通过
- web 五个 hook 的 path 字符串改 contractUrl(契约)，detail 路径参数从模板串改契约插值
  （id=null 时 enabled=false 不发请求，占位 1 只为过 positive 校验）；fetch 层与 ApiError 零改动
