+++ apps/console/src/ops/application/app-log-query.impl.service.ts
@@ -1,4 +1,4 @@
-import { type AppLogListQuery, type AppLogListResponse } from "@kagami/shared/schemas/app-log";
+import { type AppLogListQuery, type AppLogListResponse } from "@kagami/console-api/app-log";
 import type { LogDao } from "@kagami/kernel/logger/dao/log.dao";
 import type { AppLogQueryService } from "./app-log-query.service.js";
 import { mapAppLogList } from "../mappers/app-log.mapper.js";
