+++ apps/console/src/ops/application/app-log-query.service.ts
@@ -1,4 +1,4 @@
-import { type AppLogListQuery, type AppLogListResponse } from "@kagami/shared/schemas/app-log";
+import { type AppLogListQuery, type AppLogListResponse } from "@kagami/console-api/app-log";
 
 export interface AppLogQueryService {
   queryList(query: AppLogListQuery): Promise<AppLogListResponse>;
