+++ apps/console/src/ops/application/napcat-event-query.service.ts
@@ -1,7 +1,7 @@
 import {
   type NapcatEventListQuery,
   type NapcatEventListResponse,
-} from "@kagami/shared/schemas/napcat-event";
+} from "@kagami/console-api/napcat-event";
 
 export interface NapcatEventQueryService {
   queryList(query: NapcatEventListQuery): Promise<NapcatEventListResponse>;
