+++ apps/console/src/ops/application/napcat-group-message-query.impl.service.ts
@@ -1,7 +1,7 @@
 import {
   type NapcatQqMessageListQuery,
   type NapcatQqMessageListResponse,
-} from "@kagami/shared/schemas/napcat-group-message";
+} from "@kagami/console-api/napcat-group-message";
 import type { NapcatQqMessageDao } from "@kagami/persistence/dao/napcat-group-message.dao";
 import { mapNapcatQqMessageList } from "../mappers/napcat-group-message.mapper.js";
 import type { NapcatQqMessageQueryService } from "./napcat-group-message-query.service.js";
