+++ apps/console/src/ops/application/napcat-group-message-query.service.ts
@@ -1,7 +1,7 @@
 import {
   type NapcatQqMessageListQuery,
   type NapcatQqMessageListResponse,
-} from "@kagami/shared/schemas/napcat-group-message";
+} from "@kagami/console-api/napcat-group-message";
 
 export interface NapcatQqMessageQueryService {
   queryList(query: NapcatQqMessageListQuery): Promise<NapcatQqMessageListResponse>;
