+++ apps/console/src/ops/http/napcat-group-message.handler.ts
@@ -1,17 +1,14 @@
 import type { FastifyInstance } from "fastify";
-import {
-  NapcatQqMessageListQuerySchema,
-  NapcatQqMessageListResponseSchema,
-} from "@kagami/shared/schemas/napcat-group-message";
+import { registerJsonRoute } from "@kagami/http/register";
+import { consoleApiContract } from "@kagami/console-api/contract";
 import type { NapcatQqMessageQueryService } from "../application/napcat-group-message-query.service.js";
-import { registerQueryRoute } from "@kagami/http/route";
 
 type NapcatQqMessageHandlerDeps = {
   napcatQqMessageQueryService: NapcatQqMessageQueryService;
 };
 
+/** QQ 消息查询路由。路由与 schema 的单一事实源在 @kagami/console-api（#279 PR4）。 */
 export class NapcatQqMessageHandler {
-  public readonly prefix = "/napcat-group-message";
   private readonly napcatQqMessageQueryService: NapcatQqMessageQueryService;
 
   public constructor({ napcatQqMessageQueryService }: NapcatQqMessageHandlerDeps) {
@@ -19,14 +16,8 @@ export class NapcatQqMessageHandler {
   }
 
   public register(app: FastifyInstance): void {
-    registerQueryRoute({
-      app,
-      path: `${this.prefix}/query`,
-      querySchema: NapcatQqMessageListQuerySchema,
-      responseSchema: NapcatQqMessageListResponseSchema,
-      execute: ({ query }) => {
-        return this.napcatQqMessageQueryService.queryList(query);
-      },
-    });
+    registerJsonRoute(app, consoleApiContract.queryNapcatQqMessages, ({ input }) =>
+      this.napcatQqMessageQueryService.queryList(input),
+    );
   }
 }
