+++ apps/console/tsconfig.json
@@ -14,7 +14,8 @@
       "@kagami/persistence/*": ["./../../packages/persistence/src/*"],
       "@kagami/kernel/*": ["./../../packages/kernel/src/*"],
       "@kagami/http/*": ["./../../packages/http/src/*"],
-      "@kagami/shared/*": ["./../../packages/shared/src/*"]
+      "@kagami/shared/*": ["./../../packages/shared/src/*"],
+      "@kagami/console-api/*": ["./../../packages/console-api/src/*"]
     }
   },
   "include": ["src", "test"]
