+++ apps/web/src/pages/llm-history/format-status.ts
@@ -1,4 +1,4 @@
-import { type LlmChatCallStatus } from "@kagami/shared/schemas/llm-chat";
+import { type LlmChatCallStatus } from "@kagami/console-api/llm-chat-call";
 
 export function toStatusLabel(status: LlmChatCallStatus): string {
   return status === "success" ? "成功" : "失败";
