+++ apps/web/src/pages/llm-history/useLlmChatCallDetail.ts
@@ -1,7 +1,9 @@
+import { contractUrl } from "@kagami/http/url";
+import { consoleApiContract } from "@kagami/console-api/contract";
 import {
   type LlmChatCallDetailResponse,
   LlmChatCallDetailResponseSchema,
-} from "@kagami/shared/schemas/llm-chat";
+} from "@kagami/console-api/llm-chat-call";
 import { useQuery } from "@tanstack/react-query";
 import { createSchemaQueryOptions, queryKeys } from "@/lib/query";
 
@@ -12,7 +14,8 @@ export function useLlmChatCallDetail(id: number | null) {
       ReturnType<typeof queryKeys.llm.historyDetail>
     >({
       queryKey: queryKeys.llm.historyDetail(id ?? 0),
-      path: `/llm-chat-call/${id ?? 0}`,
+      // id 为 null 时 enabled=false 不会发请求；占位 1 只为通过 params 的 positive 校验。
+      path: contractUrl(consoleApiContract.getLlmChatCallDetail, { params: { id: id ?? 1 } }),
       schema: LlmChatCallDetailResponseSchema,
     }),
     enabled: id !== null,
