+++ apps/web/src/pages/napcat-event-history/NapcatEventHistoryPage.tsx
@@ -1,4 +1,4 @@
-import { type NapcatEventItem } from "@kagami/shared/schemas/napcat-event";
+import { type NapcatEventItem } from "@kagami/console-api/napcat-event";
 import { type FormEvent, useMemo } from "react";
 import { HistoryListPageLayout } from "@/components/layout/HistoryListPageLayout";
 import { Button } from "@/components/ui/button";
