+++ apps/web/src/pages/napcat-group-message-history/NapcatGroupMessageDetailPanel.tsx
@@ -1,4 +1,4 @@
-import { type NapcatQqMessageItem } from "@kagami/shared/schemas/napcat-group-message";
+import { type NapcatQqMessageItem } from "@kagami/console-api/napcat-group-message";
 import { formatOptionalDateTime } from "@/lib/format";
 import { safeStringify } from "./message-render";
