+++ packages/console-api/src/napcat-group-message.ts
